Relocated the code for generating `user_args`. Not ideal but currently I have no idea where to put it otherwise.
